# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Supercheck API",
    "version": "1.0.0",
    "description": "API for test automation, monitoring, and CI/CD integration. Supercheck provides Playwright browser testing, k6 performance testing, uptime monitoring, and alerting — all managed via API or CLI.",
    "contact": {
      "name": "Supercheck",
      "url": "https://supercheck.io"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://github.com/supercheck-io/supercheck/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://app.supercheck.io",
      "description": "Production (Cloud)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/notification-providers": {
      "get": {
        "operationId": "listNotificationProviders",
        "summary": "List notification providers",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "200": {
            "description": "List of notification providers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationProvider"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "operationId": "createNotificationProvider",
        "summary": "Create a notification provider",
        "tags": [
          "Notifications"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Slack Alerts"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "email",
                      "slack",
                      "webhook",
                      "telegram",
                      "discord",
                      "teams"
                    ]
                  },
                  "config": {
                    "type": "object",
                    "description": "Provider-specific configuration (e.g. webhookUrl for Slack)"
                  }
                },
                "required": [
                  "name",
                  "type",
                  "config"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationProvider"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/notification-providers/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "getNotificationProvider",
        "summary": "Get notification provider details",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "200": {
            "description": "Provider details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationProvider"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "operationId": "updateNotificationProvider",
        "summary": "Update a notification provider",
        "tags": [
          "Notifications"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object"
                  }
                },
                "required": [
                  "name",
                  "type",
                  "config"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationProvider"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "operationId": "deleteNotificationProvider",
        "summary": "Delete a notification provider",
        "tags": [
          "Notifications"
        ],
        "responses": {
          "200": {
            "description": "Provider deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "CLI token (`sck_live_*`) for full API access, or trigger key (`sck_trigger_*`) for job trigger endpoint only."
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or token invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Not found"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "NotificationProvider": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "email",
              "slack",
              "webhook",
              "telegram",
              "discord",
              "teams"
            ]
          },
          "enabled": {
            "type": "boolean"
          },
          "config": {
            "type": "object"
          },
          "lastUsed": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}